+++ src/Exception/RuntimeException.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\S3\Exception;
-
-final class RuntimeException extends \RuntimeException implements Exception
-{
-}