+++ src/Region.php
@@ -3,8 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\S3;
 
-use Innmind\S3\Exception\DomainException;
-use Innmind\Immutable\Str;
+use Innmind\Immutable\{
+    Str,
+    Maybe,
+};
 
 final class Region
 {
@@ -12,21 +14,30 @@ final class Region
 
     private function __construct(string $value)
     {
-        if (!Str::of($value)->matches('~^[a-z0-9\-]+$~')) {
-            throw new DomainException($value);
-        }
-
         $this->value = $value;
     }
 
     /**
      * @param literal-string $value
      *
-     * @throws DomainException
+     * @throws \DomainException
      */
     public static function of(string $value): self
     {
-        return new self($value);
+        return self::maybe($value)->match(
+            static fn($self) => $self,
+            static fn() => throw new \DomainException($value),
+        );
+    }
+
+    /**
+     * @return Maybe<self>
+     */
+    public static function maybe(string $value): Maybe
+    {
+        return Str::of($value)
+            ->maybe(static fn($value) => $value->matches('~^[a-z0-9\-]+$~'))
+            ->map(static fn($value) => new self($value->toString()));
     }
 
     public function toString(): string